ch(controllers): refactor controllers
