+++ app/controllers/api/v1/bucketlists_controller.rb
@@ -46,19 +46,6 @@ module API
       def paginate_only
         @bucketlists.paginate(params[:limit], params[:page]) if @bucketlists
       end
-
-      def user_bucketlists
-        @bucketlists = @current_user.bucketlists
-      end
-
-      def set_bucketlist
-        @bucketlist = @current_user.bucketlists.find_by(id: params[:id])
-        unless @bucketlist
-          raise(
-            ActiveRecord::RecordNotFound, Messages.not_found("bucketlist")
-          )
-        end
-      end
     end
   end
 end
