+++ app/controllers/api/v1/items_controller.rb
@@ -37,23 +37,6 @@ module API
       def paginate_items
         @items.paginate(params[:limit], params[:page]) if @items
       end
-
-      def bucketlist_items
-        @items = item_bucketlist.items if item_bucketlist
-      end
-
-      def item_bucketlist
-        @current_user.bucketlists.find_by(id: params[:bucketlist_id])
-      end
-
-      def set_bucketlist_item
-        @item = item_bucketlist.items.find_by(id: params[:id]) if item_bucketlist
-        unless @item
-          raise(
-            ActiveRecord::RecordNotFound, Messages.not_found("item")
-          )
-        end
-      end
     end
   end
 end
